+++ tests/frontends/mpd/playback_test.py
@@ -16,8 +16,18 @@ class PlaybackOptionsHandlerTest(unittest.TestCase):
         self.assertFalse(self.b.playback.consume)
         self.assert_(u'OK' in result)
 
+    def test_consume_off_without_quotes(self):
+        result = self.h.handle_request(u'consume 0')
+        self.assertFalse(self.b.playback.consume)
+        self.assert_(u'OK' in result)
+
     def test_consume_on(self):
         result = self.h.handle_request(u'consume "1"')
+        self.assertTrue(self.b.playback.consume)
+        self.assert_(u'OK' in result)
+
+    def test_consume_on_without_quotes(self):
+        result = self.h.handle_request(u'consume 1')
         self.assertTrue(self.b.playback.consume)
         self.assert_(u'OK' in result)
 
@@ -30,8 +40,18 @@ class PlaybackOptionsHandlerTest(unittest.TestCase):
         self.assertFalse(self.b.playback.random)
         self.assert_(u'OK' in result)
 
+    def test_random_off_without_quotes(self):
+        result = self.h.handle_request(u'random 0')
+        self.assertFalse(self.b.playback.random)
+        self.assert_(u'OK' in result)
+
     def test_random_on(self):
         result = self.h.handle_request(u'random "1"')
+        self.assertTrue(self.b.playback.random)
+        self.assert_(u'OK' in result)
+
+    def test_random_on_without_quotes(self):
+        result = self.h.handle_request(u'random 1')
         self.assertTrue(self.b.playback.random)
         self.assert_(u'OK' in result)
 
@@ -40,8 +60,18 @@ class PlaybackOptionsHandlerTest(unittest.TestCase):
         self.assertFalse(self.b.playback.repeat)
         self.assert_(u'OK' in result)
 
+    def test_repeat_off_without_quotes(self):
+        result = self.h.handle_request(u'repeat 0')
+        self.assertFalse(self.b.playback.repeat)
+        self.assert_(u'OK' in result)
+
     def test_repeat_on(self):
         result = self.h.handle_request(u'repeat "1"')
+        self.assertTrue(self.b.playback.repeat)
+        self.assert_(u'OK' in result)
+
+    def test_repeat_on_without_quotes(self):
+        result = self.h.handle_request(u'repeat 1')
         self.assertTrue(self.b.playback.repeat)
         self.assert_(u'OK' in result)
 
@@ -80,8 +110,18 @@ class PlaybackOptionsHandlerTest(unittest.TestCase):
         self.assertFalse(self.b.playback.single)
         self.assert_(u'OK' in result)
 
+    def test_single_off_without_quotes(self):
+        result = self.h.handle_request(u'single 0')
+        self.assertFalse(self.b.playback.single)
+        self.assert_(u'OK' in result)
+
     def test_single_on(self):
         result = self.h.handle_request(u'single "1"')
+        self.assertTrue(self.b.playback.single)
+        self.assert_(u'OK' in result)
+
+    def test_single_on_without_quotes(self):
+        result = self.h.handle_request(u'single 1')
         self.assertTrue(self.b.playback.single)
         self.assert_(u'OK' in result)
 
@@ -136,8 +176,7 @@ class PlaybackControlHandlerTest(unittest.TestCase):
         self.assert_(u'OK' in result)
 
     def test_pause_off(self):
-        track = Track()
-        self.b.current_playlist.load([track])
+        self.b.current_playlist.load([Track()])
         self.h.handle_request(u'play "0"')
         self.h.handle_request(u'pause "1"')
         result = self.h.handle_request(u'pause "0"')
@@ -145,16 +184,26 @@ class PlaybackControlHandlerTest(unittest.TestCase):
         self.assertEqual(self.b.playback.PLAYING, self.b.playback.state)
 
     def test_pause_on(self):
-        track = Track()
-        self.b.current_playlist.load([track])
+        self.b.current_playlist.load([Track()])
         self.h.handle_request(u'play "0"')
         result = self.h.handle_request(u'pause "1"')
         self.assert_(u'OK' in result)
         self.assertEqual(self.b.playback.PAUSED, self.b.playback.state)
 
+    def test_pause_toggle(self):
+        self.b.current_playlist.load([Track()])
+        result = self.h.handle_request(u'play "0"')
+        self.assert_(u'OK' in result)
+        self.assertEqual(self.b.playback.PLAYING, self.b.playback.state)
+        result = self.h.handle_request(u'pause')
+        self.assert_(u'OK' in result)
+        self.assertEqual(self.b.playback.PAUSED, self.b.playback.state)
+        result = self.h.handle_request(u'pause')
+        self.assert_(u'OK' in result)
+        self.assertEqual(self.b.playback.PLAYING, self.b.playback.state)
+
     def test_play_without_pos(self):
-        track = Track()
-        self.b.current_playlist.load([track])
+        self.b.current_playlist.load([Track()])
         self.b.playback.state = self.b.playback.PAUSED
         result = self.h.handle_request(u'play')
         self.assert_(u'OK' in result)
@@ -163,6 +212,12 @@ class PlaybackControlHandlerTest(unittest.TestCase):
     def test_play_with_pos(self):
         self.b.current_playlist.load([Track()])
         result = self.h.handle_request(u'play "0"')
+        self.assert_(u'OK' in result)
+        self.assertEqual(self.b.playback.PLAYING, self.b.playback.state)
+
+    def test_play_with_pos_without_quotes(self):
+        self.b.current_playlist.load([Track()])
+        result = self.h.handle_request(u'play 0')
         self.assert_(u'OK' in result)
         self.assertEqual(self.b.playback.PLAYING, self.b.playback.state)